+++ packages/std/src/query/mod.rs
@@ -5,6 +5,25 @@ use serde::{Deserialize, Serialize};
 use crate::Binary;
 use crate::Empty;
 
+/// Implements a hidden constructor for query responses.
+macro_rules! impl_response_constructor {
+    ( $response:ty, $( $field: ident : $t: ty),* ) => {
+        impl $response {
+            /// Constructor for testing frameworks such as cw-multi-test.
+            /// This is required because query response types should be #[non_exhaustive].
+            /// As a contract developer you should not need this constructor since
+            /// query responses are constructed for you via deserialization.
+            ///
+            /// Warning: This can change in breaking ways in minor versions.
+            #[doc(hidden)]
+            #[allow(dead_code)]
+            pub fn new($( $field: $t),*) -> Self {
+                Self { $( $field ),* }
+            }
+        }
+    };
+}
+
 mod bank;
 mod distribution;
 mod ibc;
@@ -12,23 +31,11 @@ mod query_response;
 mod staking;
 mod wasm;
 
-#[cfg(feature = "cosmwasm_1_1")]
-pub use bank::SupplyResponse;
-pub use bank::{AllBalanceResponse, BalanceResponse, BankQuery};
-#[cfg(feature = "cosmwasm_1_3")]
-pub use bank::{AllDenomMetadataResponse, DenomMetadataResponse};
-#[cfg(feature = "cosmwasm_1_3")]
-pub use distribution::{DelegatorWithdrawAddressResponse, DistributionQuery};
-#[cfg(feature = "stargate")]
-pub use ibc::{ChannelResponse, IbcQuery, ListChannelsResponse, PortIdResponse};
-#[cfg(feature = "staking")]
-pub use staking::{
-    AllDelegationsResponse, AllValidatorsResponse, BondedDenomResponse, Delegation,
-    DelegationResponse, FullDelegation, StakingQuery, Validator, ValidatorResponse,
-};
-#[cfg(feature = "cosmwasm_1_2")]
-pub use wasm::CodeInfoResponse;
-pub use wasm::{ContractInfoResponse, WasmQuery};
+pub use bank::*;
+pub use distribution::*;
+pub use ibc::*;
+pub use staking::*;
+pub use wasm::*;
 
 #[non_exhaustive]
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]