+++ packages/std/src/query/staking.rs
@@ -1,9 +1,9 @@
-#![cfg(feature = "staking")]
-
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
 use crate::{Addr, Coin, Decimal};
+
+use super::query_response::QueryResponseType;
 
 #[non_exhaustive]
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
@@ -40,12 +40,20 @@ pub struct BondedDenomResponse {
     pub denom: String,
 }
 
+impl QueryResponseType for BondedDenomResponse {}
+
+impl_response_constructor!(BondedDenomResponse, denom: String);
+
 /// DelegationsResponse is data format returned from StakingRequest::AllDelegations query
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
 pub struct AllDelegationsResponse {
     pub delegations: Vec<Delegation>,
 }
+
+impl QueryResponseType for AllDelegationsResponse {}
+
+impl_response_constructor!(AllDelegationsResponse, delegations: Vec<Delegation>);
 
 /// Delegation is basic (cheap to query) data about a delegation.
 ///
@@ -76,6 +84,10 @@ pub struct DelegationResponse {
     pub delegation: Option<FullDelegation>,
 }
 
+impl QueryResponseType for DelegationResponse {}
+
+impl_response_constructor!(DelegationResponse, delegation: Option<FullDelegation>);
+
 /// FullDelegation is all the info on the delegation, some (like accumulated_reward and can_redelegate)
 /// is expensive to query.
 ///
@@ -101,11 +113,19 @@ pub struct AllValidatorsResponse {
     pub validators: Vec<Validator>,
 }
 
+impl QueryResponseType for AllValidatorsResponse {}
+
+impl_response_constructor!(AllValidatorsResponse, validators: Vec<Validator>);
+
 /// The data format returned from StakingRequest::Validator query
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 pub struct ValidatorResponse {
     pub validator: Option<Validator>,
 }
+
+impl QueryResponseType for ValidatorResponse {}
+
+impl_response_constructor!(ValidatorResponse, validator: Option<Validator>);
 
 /// Instances are created in the querier.
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]