+++ packages/vm/src/cache.rs
@@ -4,7 +4,7 @@ use std::io::{Read, Write};
 use std::marker::PhantomData;
 use std::path::{Path, PathBuf};
 use std::sync::Mutex;
-use wasmer::{Engine, NativeEngineExt};
+use wasmer::{Engine, Store};
 
 use crate::backend::{Backend, BackendApi, Querier, Storage};
 use crate::capabilities::required_capabilities_from_module;
@@ -14,9 +14,10 @@ use crate::errors::{VmError, VmResult};
 use crate::filesystem::mkdir_p;
 use crate::instance::{Instance, InstanceOptions};
 use crate::modules::{CachedModule, FileSystemCache, InMemoryCache, PinnedMemoryCache};
+use crate::parsed_wasm::ParsedWasm;
 use crate::size::Size;
-use crate::static_analysis::{deserialize_wasm, has_ibc_entry_points};
-use crate::wasm_backend::{compile, make_store_with_engine};
+use crate::static_analysis::has_ibc_entry_points;
+use crate::wasm_backend::{compile, make_compiling_engine, make_runtime_engine};
 
 const STATE_DIR: &str = "state";
 // Things related to the state of the blockchain.
@@ -67,13 +68,21 @@ pub struct CacheOptions {
 pub struct CacheInner {
     /// The directory in which the Wasm blobs are stored in the file system.
     wasm_path: PathBuf,
-    /// Instances memory limit in bytes. Use a value that is divisible by the Wasm page size 65536,
-    /// e.g. full MiBs.
-    instance_memory_limit: Size,
     pinned_memory_cache: PinnedMemoryCache,
     memory_cache: InMemoryCache,
     fs_cache: FileSystemCache,
     stats: Stats,
+    /// A single engine to execute all contracts in this cache instance (usually
+    /// this means all contracts in the process).
+    ///
+    /// This engine is headless, i.e. does not contain a Singlepass compiler.
+    /// It only executes modules compiled with other engines.
+    ///
+    /// The engine has one memory limit set which is the same for all contracts
+    /// running with it. If different memory limits would be needed for different
+    /// contracts at some point, we'd need multiple engines. This is because the tunables
+    /// that control the limit are attached to the engine.
+    runtime_engine: Engine,
 }
 
 pub struct Cache<A: BackendApi, S: Storage, Q: Querier> {
@@ -132,11 +141,11 @@ where
             available_capabilities,
             inner: Mutex::new(CacheInner {
                 wasm_path,
-                instance_memory_limit,
                 pinned_memory_cache: PinnedMemoryCache::new(),
                 memory_cache: InMemoryCache::new(memory_cache_size),
                 fs_cache,
                 stats: Stats::default(),
+                runtime_engine: make_runtime_engine(Some(instance_memory_limit)),
             }),
             type_storage: PhantomData::<S>,
             type_api: PhantomData::<A>,
@@ -191,7 +200,9 @@ where
     /// When a Wasm blob is stored which was previously checked (e.g. as part of state sync),
     /// use this function.
     pub fn save_wasm_unchecked(&self, wasm: &[u8]) -> VmResult<Checksum> {
-        let (_engine, module) = compile(wasm, &[])?;
+        // We need a new engine for each Wasm -> module compilation due to the metering middleware.
+        let compiling_engine = make_compiling_engine(None);
+        let module = compile(&compiling_engine, wasm)?;
 
         let mut cache = self.inner.lock().unwrap();
         let checksum = save_wasm_to_disk(&cache.wasm_path, wasm)?;
@@ -244,7 +255,7 @@ where
     pub fn analyze(&self, checksum: &Checksum) -> VmResult<AnalysisReport> {
         // Here we could use a streaming deserializer to slightly improve performance. However, this way it is DRYer.
         let wasm = self.load_wasm(checksum)?;
-        let module = deserialize_wasm(&wasm)?;
+        let module = ParsedWasm::parse(&wasm)?;
         Ok(AnalysisReport {
             has_ibc_entry_points: has_ibc_entry_points(&module),
             required_capabilities: required_capabilities_from_module(&module),
@@ -270,23 +281,24 @@ where
         // for a not-so-relevant use case.
 
         // Try to get module from file system cache
-        let engine = Engine::headless();
-        if let Some((module, module_size)) = cache.fs_cache.load(checksum, &engine)? {
+        if let Some((module, module_size)) = cache.fs_cache.load(checksum, &cache.runtime_engine)? {
             cache.stats.hits_fs_cache = cache.stats.hits_fs_cache.saturating_add(1);
             return cache
                 .pinned_memory_cache
-                .store(checksum, (engine, module), module_size);
+                .store(checksum, module, module_size);
         }
 
         // Re-compile from original Wasm bytecode
-        let code = self.load_wasm_with_path(&cache.wasm_path, checksum)?;
+        let wasm = self.load_wasm_with_path(&cache.wasm_path, checksum)?;
         cache.stats.misses = cache.stats.misses.saturating_add(1);
-        let (engine, module) = compile(&code, &[])?;
+        // Module will run with a different engine, so we can set memory limit to None
+        let engine = make_compiling_engine(None);
+        let module = compile(&engine, &wasm)?;
         // Store into the fs cache too
         let module_size = cache.fs_cache.store(checksum, &module)?;
         cache
             .pinned_memory_cache
-            .store(checksum, (engine, module), module_size)
+            .store(checksum, module, module_size)
     }
 
     /// Unpins a Module, i.e. removes it from the pinned memory cache.
@@ -310,8 +322,7 @@ where
         backend: Backend<A, S, Q>,
         options: InstanceOptions,
     ) -> VmResult<Instance<A, S, Q>> {
-        let (cached, memory_limit, _from_pinned) = self.get_module(checksum)?;
-        let store = make_store_with_engine(cached.engine, Some(memory_limit));
+        let (cached, store) = self.get_module(checksum)?;
         let instance = Instance::from_module(
             store,
             &cached.module,
@@ -327,35 +338,36 @@ where
     /// Returns a module tied to a previously saved Wasm.
     /// Depending on availability, this is either generated from a memory cache, file system cache or Wasm code.
     /// This is part of `get_instance` but pulled out to reduce the locking time.
-    fn get_module(&self, checksum: &Checksum) -> VmResult<(CachedModule, Size, bool)> {
+    fn get_module(&self, checksum: &Checksum) -> VmResult<(CachedModule, Store)> {
         let mut cache = self.inner.lock().unwrap();
         // Try to get module from the pinned memory cache
         if let Some(element) = cache.pinned_memory_cache.load(checksum)? {
             cache.stats.hits_pinned_memory_cache =
                 cache.stats.hits_pinned_memory_cache.saturating_add(1);
-            return Ok((element, cache.instance_memory_limit, true));
+            let store = Store::new(cache.runtime_engine.clone());
+            return Ok((element, store));
         }
 
         // Get module from memory cache
         if let Some(element) = cache.memory_cache.load(checksum)? {
             cache.stats.hits_memory_cache = cache.stats.hits_memory_cache.saturating_add(1);
-            return Ok((element, cache.instance_memory_limit, false));
+            let store = Store::new(cache.runtime_engine.clone());
+            return Ok((element, store));
         }
 
         // Get module from file system cache
-        let engine = Engine::headless();
-        if let Some((module, module_size)) = cache.fs_cache.load(checksum, &engine)? {
+        if let Some((module, module_size)) = cache.fs_cache.load(checksum, &cache.runtime_engine)? {
             cache.stats.hits_fs_cache = cache.stats.hits_fs_cache.saturating_add(1);
 
             cache
                 .memory_cache
-                .store(checksum, (engine.clone(), module.clone()), module_size)?;
+                .store(checksum, module.clone(), module_size)?;
             let cached = CachedModule {
-                engine,
                 module,
-                size: module_size,
+                size_estimate: module_size,
             };
-            return Ok((cached, cache.instance_memory_limit, false));
+            let store = Store::new(cache.runtime_engine.clone());
+            return Ok((cached, store));
         }
 
         // Re-compile module from wasm
@@ -365,18 +377,20 @@ where
         // stored the old module format.
         let wasm = self.load_wasm_with_path(&cache.wasm_path, checksum)?;
         cache.stats.misses = cache.stats.misses.saturating_add(1);
-        let (engine, module) = compile(&wasm, &[])?;
+        // Module will run with a different engine, so we can set memory limit to None
+        let engine = make_compiling_engine(None);
+        let module = compile(&engine, &wasm)?;
         let module_size = cache.fs_cache.store(checksum, &module)?;
 
         cache
             .memory_cache
-            .store(checksum, (engine.clone(), module.clone()), module_size)?;
+            .store(checksum, module.clone(), module_size)?;
         let cached = CachedModule {
-            engine,
             module,
-            size: module_size,
+            size_estimate: module_size,
         };
-        Ok((cached, cache.instance_memory_limit, false))
+        let store = Store::new(cache.runtime_engine.clone());
+        Ok((cached, store))
     }
 }
 
@@ -558,7 +572,7 @@ mod tests {
         let save_result = cache.save_wasm(&wasm);
         match save_result.unwrap_err() {
             VmError::StaticValidationErr { msg, .. } => {
-                assert_eq!(msg, "Wasm contract doesn\'t have a memory section")
+                assert_eq!(msg, "Wasm contract must contain exactly one memory")
             }
             e => panic!("Unexpected error {e:?}"),
         }