+++ packages/vm/src/imports.rs
@@ -583,7 +583,7 @@ mod tests {
     use crate::backend::{BackendError, Storage};
     use crate::size::Size;
     use crate::testing::{MockApi, MockQuerier, MockStorage};
-    use crate::wasm_backend::{compile, make_store_with_engine};
+    use crate::wasm_backend::{compile, make_compiling_engine};
 
     static CONTRACT: &[u8] = include_bytes!("../testdata/hackatom.wasm");
 
@@ -620,8 +620,9 @@ mod tests {
         let gas_limit = TESTING_GAS_LIMIT;
         let env = Environment::new(api, gas_limit);
 
-        let (engine, module) = compile(CONTRACT, &[]).unwrap();
-        let mut store = make_store_with_engine(engine, TESTING_MEMORY_LIMIT);
+        let engine = make_compiling_engine(TESTING_MEMORY_LIMIT);
+        let module = compile(&engine, CONTRACT).unwrap();
+        let mut store = Store::new(engine);
 
         let fe = FunctionEnv::new(&mut store, env);