+++ packages/vm/src/modules/pinned_memory_cache.rs
@@ -1,5 +1,5 @@
 use std::collections::HashMap;
-use wasmer::{Engine, Module};
+use wasmer::Module;
 
 use super::cached_module::CachedModule;
 use crate::{Checksum, VmResult};
@@ -20,15 +20,14 @@ impl PinnedMemoryCache {
     pub fn store(
         &mut self,
         checksum: &Checksum,
-        element: (Engine, Module),
-        size: usize,
+        element: Module,
+        module_size: usize,
     ) -> VmResult<()> {
         self.modules.insert(
             *checksum,
             CachedModule {
-                engine: element.0,
-                module: element.1,
-                size,
+                module: element,
+                size_estimate: module_size,
             },
         );
         Ok(())
@@ -64,17 +63,24 @@ impl PinnedMemoryCache {
     /// This is based on the values provided with `store`. No actual
     /// memory size is measured here.
     pub fn size(&self) -> usize {
-        self.modules.values().map(|module| module.size).sum()
+        self.modules
+            .iter()
+            .map(|(key, module)| std::mem::size_of_val(key) + module.size_estimate)
+            .sum()
     }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::wasm_backend::{compile, make_store_with_engine};
-    use wasmer::{imports, Instance as WasmerInstance};
+    use crate::{
+        wasm_backend::{compile, make_compiling_engine},
+        Size,
+    };
+    use wasmer::{imports, Instance as WasmerInstance, Store};
     use wasmer_middlewares::metering::set_remaining_points;
 
+    const TESTING_MEMORY_LIMIT: Option<Size> = Some(Size::mebi(16));
     const TESTING_GAS_LIMIT: u64 = 500_000_000;
 
     #[test]
@@ -99,11 +105,12 @@ mod tests {
         assert!(cache_entry.is_none());
 
         // Compile module
-        let (engine, original) = compile(&wasm, &[]).unwrap();
-        let mut store = make_store_with_engine(engine.clone(), None);
+        let engine = make_compiling_engine(TESTING_MEMORY_LIMIT);
+        let original = compile(&engine, &wasm).unwrap();
 
         // Ensure original module can be executed
         {
+            let mut store = Store::new(engine.clone());
             let instance = WasmerInstance::new(&mut store, &original, &imports! {}).unwrap();
             set_remaining_points(&mut store, &instance, TESTING_GAS_LIMIT);
             let add_one = instance.exports.get_function("add_one").unwrap();
@@ -112,14 +119,14 @@ mod tests {
         }
 
         // Store module
-        cache.store(&checksum, (engine, original), 0).unwrap();
+        cache.store(&checksum, original, 0).unwrap();
 
         // Load module
         let cached = cache.load(&checksum).unwrap().unwrap();
-        let mut store = make_store_with_engine(cached.engine, None);
 
         // Ensure cached module can be executed
         {
+            let mut store = Store::new(engine);
             let instance = WasmerInstance::new(&mut store, &cached.module, &imports! {}).unwrap();
             set_remaining_points(&mut store, &instance, TESTING_GAS_LIMIT);
             let add_one = instance.exports.get_function("add_one").unwrap();
@@ -148,8 +155,9 @@ mod tests {
         assert!(!cache.has(&checksum));
 
         // Add
-        let (engine, original) = compile(&wasm, &[]).unwrap();
-        cache.store(&checksum, (engine, original), 0).unwrap();
+        let engine = make_compiling_engine(TESTING_MEMORY_LIMIT);
+        let original = compile(&engine, &wasm).unwrap();
+        cache.store(&checksum, original, 0).unwrap();
 
         assert!(cache.has(&checksum));
 
@@ -179,8 +187,9 @@ mod tests {
         assert_eq!(cache.len(), 0);
 
         // Add
-        let (engine, original) = compile(&wasm, &[]).unwrap();
-        cache.store(&checksum, (engine, original), 0).unwrap();
+        let engine = make_compiling_engine(TESTING_MEMORY_LIMIT);
+        let original = compile(&engine, &wasm).unwrap();
+        cache.store(&checksum, original, 0).unwrap();
 
         assert_eq!(cache.len(), 1);
 
@@ -221,20 +230,20 @@ mod tests {
         assert_eq!(cache.size(), 0);
 
         // Add 1
-        cache
-            .store(&checksum1, compile(&wasm1, &[]).unwrap(), 500)
-            .unwrap();
-        assert_eq!(cache.size(), 500);
+        let engine1 = make_compiling_engine(TESTING_MEMORY_LIMIT);
+        let module = compile(&engine1, &wasm1).unwrap();
+        cache.store(&checksum1, module, 500).unwrap();
+        assert_eq!(cache.size(), 532);
 
         // Add 2
-        cache
-            .store(&checksum2, compile(&wasm2, &[]).unwrap(), 300)
-            .unwrap();
-        assert_eq!(cache.size(), 800);
+        let engine2 = make_compiling_engine(TESTING_MEMORY_LIMIT);
+        let module = compile(&engine2, &wasm2).unwrap();
+        cache.store(&checksum2, module, 300).unwrap();
+        assert_eq!(cache.size(), 532 + 332);
 
         // Remove 1
         cache.remove(&checksum1).unwrap();
-        assert_eq!(cache.size(), 300);
+        assert_eq!(cache.size(), 332);
 
         // Remove 2
         cache.remove(&checksum2).unwrap();